clip_finetune/requirements.txt
torch
torchvision
open_clip_torch
pillow
numpy
tqdm
